requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML >= 5.4.1
mysqlclient >= 2.0.3
pyserial >= 3.5